+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 ## [Unreleased]
 ### Added
 - e-Bus favicon
+- CO2 in gram on display route page
+
+### Fixed
+- Answer shows up shortly before answer score animation
 
 ## [0.9.0] - 2020-05-19
 ### Added